requirements.txt
boto3~=1.22.4
botocore
python-daemon
rich
